+++ Source/cmFileLock.h
@@ -37,7 +37,9 @@ class cmFileLock
     * @brief Lock the file.
     * @param timeoutSec Lock timeout. If -1 try until success or fatal error.
     */
-  cmFileLockResult Lock(const std::string& filename, unsigned timeoutSec);
+  cmFileLockResult Lock(
+      const std::string& filename, unsigned long long timeoutSec
+  );
 
   /**
     * @brief Unlock the file.
@@ -57,7 +59,7 @@ class cmFileLock
 
   cmFileLockResult OpenFile();
   cmFileLockResult LockWithoutTimeout();
-  cmFileLockResult LockWithTimeout(unsigned timeoutSec);
+  cmFileLockResult LockWithTimeout(unsigned long long timeoutSec);
 
 #if defined(_WIN32)
   typedef HANDLE FileId;
